Renaming frozen transforms to zero transforms
